Move add/edit dialog to own component
