+++ src/components/pages/Participants.tsx
@@ -4,7 +4,7 @@ import Toolbar from '@material-ui/core/Toolbar';
 import Typography from '@material-ui/core/Typography';
 import Paper from '@material-ui/core/Paper';
 import { createStyles, Theme, withStyles, WithStyles } from '@material-ui/core/styles';
-import { TableCell, TableRow, TableHead, Table, TableBody, Grid, Button, Tooltip, IconButton, Hidden, Icon, FormGroup } from '@material-ui/core';
+import { TableCell, TableRow, TableHead, Table, TableBody, Grid, Button, Tooltip, IconButton, Hidden, Icon } from '@material-ui/core';
 import RefreshIcon from '@material-ui/icons/Refresh';
 import PeopleIcon from '@material-ui/icons/People';
 import EditIcon from '@material-ui/icons/Edit';
@@ -13,15 +13,7 @@ import NotInterestedIcon from '@material-ui/icons/NotInterested';
 import CheckIcon from '@material-ui/icons/Check';
 import { IParticipant } from '../../data/Model';
 import { participantRepository } from '../../data/Repository';
-import Dialog from '@material-ui/core/Dialog';
-import DialogActions from '@material-ui/core/DialogActions';
-import DialogContent from '@material-ui/core/DialogContent';
-import DialogTitle from '@material-ui/core/DialogTitle';
-import Checkbox from '@material-ui/core/Checkbox';
-import FormControlLabel from '@material-ui/core/FormControlLabel';
-import FormControl from '@material-ui/core/FormControl';
-import Input from '@material-ui/core/Input';
-import InputLabel from '@material-ui/core/InputLabel';
+import ParticipantDialog from '../ParticipantDialog';
 
 const styles = (theme: Theme) =>
   createStyles({
@@ -47,15 +39,12 @@ const styles = (theme: Theme) =>
     },
   });
 
-interface IParticipantProps extends WithStyles<typeof styles> {}
+interface IParticipantProps extends WithStyles<typeof styles> {
+}
 
 interface IParticipantState {
   participants: IParticipant[];
   dialogOpen: boolean;
-  dialogFirstName: string;
-  dialogLastName: string;
-  dialogBirthdate: Date;
-  dialogHasPassport: boolean;
   dialogIsEdit: boolean;
   dialogEditId?: string;
 }
@@ -63,14 +52,10 @@ interface IParticipantState {
 class Participants extends Component<IParticipantProps, IParticipantState> {
   constructor(props: IParticipantProps) {
     super(props);
-    
+
     this.state = {
       participants: [],
       dialogOpen: false,
-      dialogBirthdate: new Date(),
-      dialogFirstName: "",
-      dialogHasPassport: false,
-      dialogLastName: "",
       dialogIsEdit: false
     }
   } 
@@ -87,46 +72,21 @@ class Participants extends Component<IParticipantProps, IParticipantState> {
   addActionClick = () => {
     this.setState({
       dialogOpen: true,
-      dialogBirthdate: new Date(),
-      dialogFirstName: "",
-      dialogHasPassport: false,
-      dialogLastName: "",
       dialogIsEdit: false
     });
   }
 
-  handleClose = () => {
-    this.setState({ dialogOpen: false })
+  handleParticipantAdded = (participant: IParticipant) => {
+    this.setState({ 
+      dialogOpen: false,
+      participants: this.state.participants.concat(participant)
+    })
   }
 
-  handleSave = () => {
-    if (this.state.dialogIsEdit && this.state.dialogEditId) {
-      participantRepository.get(this.state.dialogEditId)
-        .then((p) => {
-          p.birthdate = this.state.dialogBirthdate.toISOString();
-          p.firstName = this.state.dialogFirstName;
-          p.lastName = this.state.dialogLastName;
-          p.hasPassport = this.state.dialogHasPassport;
-
-          participantRepository.update(p)
-            .then(() => {
-              this.setState({ dialogOpen: false, participants:this.state.participants.map((item, i) => { return (item._id === p._id) ? p : item; }) })
-            })
-        })
-    } else {
-      let participant: IParticipant = {
-        _id: '',
-        firstName: this.state.dialogFirstName,
-        lastName: this.state.dialogLastName,
-        birthdate: this.state.dialogBirthdate.toISOString(),
-        hasPassport: this.state.dialogHasPassport
-      }
-
-      participantRepository.add(participant)
-        .then(() => {
-          this.setState({ dialogOpen: false, participants: this.state.participants.concat(participant) })
-        });
-    }
+  handleParticipantUpdated = (participant: IParticipant) => {
+    this.setState({ 
+      dialogOpen: false,
+      participants:this.state.participants.map((item, i) => { return (item._id === participant._id) ? participant : item; }) })
   }
 
   editActionClick = (rowId: string, e: React.MouseEvent<HTMLElement>) => {
@@ -140,10 +100,6 @@ class Participants extends Component<IParticipantProps, IParticipantState> {
 
     this.setState({
       dialogOpen: true,
-      dialogBirthdate: new Date(participant.birthdate),
-      dialogFirstName: participant.firstName,
-      dialogHasPassport: participant.hasPassport,
-      dialogLastName: participant.lastName,
       dialogIsEdit: true,
       dialogEditId: rowId
     });
@@ -231,37 +187,7 @@ class Participants extends Component<IParticipantProps, IParticipantState> {
                   </TableBody>
                 </Table>
               </div>
-              <Dialog open={this.state.dialogOpen} onClose={this.handleClose} fullWidth={true} maxWidth={"xs"}>
-                <DialogTitle id="form-dialog-title">{ this.state.dialogIsEdit ? "Edit Participant" : "Add Participant"}</DialogTitle>
-                <DialogContent>
-                  <FormGroup>
-                    <FormControl>
-                      <InputLabel htmlFor="firstname">First Name</InputLabel>
-                      <Input id="firstname" autoFocus defaultValue={this.state.dialogFirstName} onChange={(e) => this.setState({ dialogFirstName: e.target.value })} />
-                    </FormControl>
-                    <FormControl>
-                      <InputLabel htmlFor="lastname">Last Name</InputLabel>
-                      <Input id="lastname" defaultValue={this.state.dialogLastName} onChange={(e) => this.setState({ dialogLastName: e.target.value })}/>
-                    </FormControl>
-                    <FormControl>
-                      <InputLabel htmlFor="birthdate" shrink>Birthday</InputLabel>
-                      <Input id="birthdate" type="date" defaultValue={this.state.dialogBirthdate.toISOString().substr(0,10)} onChange={(e) => this.setState({ dialogBirthdate: new Date(e.target.value) })}/>
-                    </FormControl>
-                  <FormControlLabel
-                      control={<Checkbox value="jason" defaultChecked={this.state.dialogHasPassport} onChange={(e) => this.setState({ dialogHasPassport: e.target.checked })} />}
-                      label="Passport available"
-                    />
-                  </FormGroup>
-                </DialogContent>
-                <DialogActions>
-                  <Button onClick={this.handleClose} color="primary">
-                    Cancel
-                  </Button>
-                  <Button onClick={this.handleSave} color="primary">
-                    Save
-                  </Button>
-                </DialogActions>
-              </Dialog>
+              <ParticipantDialog open={this.state.dialogOpen} editMode={this.state.dialogIsEdit} editParticipant={this.state.dialogEditId} onAdded={this.handleParticipantAdded} onUpdated={this.handleParticipantUpdated} />
               </Paper>
     )
   };
